Corregido

+++ install.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # CPacket Installer
-# Version 2.1
+# Version 2.1 for Debian/Ubuntu
 # Author: Juanjo Salvador
 # Website: http://juanjosalvador.github.io/
 # Repo: https://github.com/JuanjoSalvador/CPacket
@@ -12,11 +12,18 @@
 		exit
 
 	else
-
+		
 		echo "Instalando CPacket..."
-
-		cp ./cpacket /usr/local/bin
-
-		echo "Finalizado"
-
+		echo "CPacket depende de zip."
+		
+		if [ -e "/usr/bin/zip" ]; then
+			echo "El paquete zip está instalado."
+			cp ./cpacket /usr/local/bin
+			echo "Finalizado"
+		else
+			echo "El paquete zip no está instalado. Instalando..."
+			apt-get install zip
+			cp ./cpacket /usr/local/bin
+			echo "Finalizado"
+		fi
 	fi
